Added the start of a Metaschema Metaschema implementation by adding tests around parsing a Metaschema Metaschema. The full metaschema won't be parsed until a solution to https://github.com/usnistgov/metaschema/issues/228 is implemented. Added a new databind-metaschema/pom.xml Maven module as a placeholder for new parsing code based on generated Metaschema classes.
