+++ databind-metaschema/pom.xml
@@ -1,0 +1,176 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>gov.nist.secauto.metaschema</groupId>
+		<artifactId>metaschema-framework</artifactId>
+		<version>1.0.0-M2-SNAPSHOT</version>
+	</parent>
+
+	<artifactId>metaschema-databind-metaschema</artifactId>
+
+	<name>Metaschema Format Databinding</name>
+
+	<url>${site.url}${project.artifactId}/</url>
+	<distributionManagement>
+		<site>
+			<id>nist-pages</id>
+			<url>${site.url}${project.artifactId}/</url>
+		</site>
+	</distributionManagement>
+
+	<description>Code to read and write Metaschema modules and extensions.</description>
+
+	<dependencies>
+		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>metaschema-schema-generator</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>metaschema-databind</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>com.google.auto.service</groupId>
+			<artifactId>auto-service-annotations</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>com.github.spotbugs</groupId>
+			<artifactId>spotbugs-annotations</artifactId>
+		</dependency>
+	</dependencies>
+
+	<build>
+		<resources>
+			<resource>
+				<directory>src/main/resources</directory>
+			</resource>
+			<resource>
+				<directory>src/main/distro</directory>
+				<filtering>true</filtering>
+				<targetPath>${project.build.directory}/generated-distro</targetPath>
+			</resource>
+			<resource>
+				<directory>${project.build.directory}/generated-resources/license</directory>
+			</resource>
+		</resources>
+		<pluginManagement>
+			<plugins>
+				<plugin>
+					<groupId>org.apache.maven.plugins</groupId>
+					<artifactId>maven-pmd-plugin</artifactId>
+					<configuration>
+						<excludeRoots>
+							<excludeRoot>target/generated-sources/metaschema</excludeRoot>
+						</excludeRoots>
+					</configuration>
+				</plugin>
+				<plugin>
+					<groupId>org.apache.maven.plugins</groupId>
+					<artifactId>maven-checkstyle-plugin</artifactId>
+					<configuration>
+						<excludes>
+							**/gov/nist/secauto/metaschema/metaschema/model/**/*</excludes>
+					</configuration>
+				</plugin>
+			</plugins>
+		</pluginManagement>
+		<plugins>
+			<plugin>
+				<groupId>gov.nist.secauto.metaschema</groupId>
+				<artifactId>metaschema-maven-plugin</artifactId>
+				<version>${project.version}</version>
+				<executions>
+					<execution>
+						<id>metaschema-codegen</id>
+						<phase>generate-sources</phase>
+						<goals>
+							<goal>generate-sources</goal>
+						</goals>
+						<configuration>
+							<metaschemaDir>${project.basedir}/../core/metaschema/schema/metaschema</metaschemaDir>
+							<outputDirectory>${project.build.directory}/generated-sources/metaschema</outputDirectory>
+							<configs>
+								<config>
+									${project.basedir}/src/main/metaschema-bindings/metaschema-metaschema-bindings.xml</config>
+							</configs>
+							<includes>
+								<include>metaschema-metaschema.xml</include>
+							</includes>
+						</configuration>
+					</execution>
+					<execution>
+						<id>generate-xml-schemas</id>
+						<goals>
+							<goal>generate-schemas</goal>
+						</goals>
+						<configuration>
+							<metaschemaDir>${project.basedir}/../core/metaschema/schema/metaschema</metaschemaDir>
+							<outputDirectory>${project.build.directory}/generated-resources/metaschema/schema/xml</outputDirectory>
+							<includes>
+								<include>metaschema-metaschema.xml</include>
+							</includes>
+							<formats>
+								<format>xsd</format>
+							</formats>
+						</configuration>
+					</execution>
+					<execution>
+						<id>generate-json-schemas</id>
+						<goals>
+							<goal>generate-schemas</goal>
+						</goals>
+						<configuration>
+							<metaschemaDir>${project.basedir}/../core/metaschema/schema/metaschema</metaschemaDir>
+							<outputDirectory>${project.build.directory}/generated-resources/oscal/schema/json</outputDirectory>
+							<includes>
+								<include>metaschema-metaschema.xml</include>
+							</includes>
+							<formats>
+								<format>json</format>
+							</formats>
+						</configuration>
+					</execution>
+				</executions>
+			</plugin>
+			<plugin>
+				<groupId>org.codehaus.mojo</groupId>
+				<artifactId>build-helper-maven-plugin</artifactId>
+				<executions>
+					<execution>
+						<id>add-source</id>
+						<phase>generate-sources</phase>
+						<goals>
+							<goal>add-source</goal>
+						</goals>
+						<configuration>
+							<sources>
+								<source>${project.build.directory}/generated-sources/metaschema</source>
+							</sources>
+						</configuration>
+					</execution>
+					<execution>
+						<id>add-metaschema-resources</id>
+						<phase>generate-resources</phase>
+						<goals>
+							<goal>add-resource</goal>
+						</goals>
+						<configuration>
+							<resources>
+								<resource>
+									<directory>${project.build.directory}/generated-resources/metaschema</directory>
+									<includes>
+										<include>**/*.xsd</include>
+										<include>**/*.json</include>
+									</includes>
+								</resource>
+							</resources>
+						</configuration>
+					</execution>
+				</executions>
+			</plugin>
+        </plugins>
+	</build>
+</project>
