+++ databind/src/test/java/gov/nist/secauto/metaschema/databind/codegen/BasicMetaschemaTest.java
@@ -26,142 +26,24 @@
 
 package gov.nist.secauto.metaschema.databind.codegen;
 
-import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import gov.nist.secauto.metaschema.core.model.IModule;
 import gov.nist.secauto.metaschema.core.model.MetaschemaException;
-import gov.nist.secauto.metaschema.core.model.xml.ModuleLoader;
 import gov.nist.secauto.metaschema.core.util.ObjectUtils;
-import gov.nist.secauto.metaschema.databind.IBindingContext;
-import gov.nist.secauto.metaschema.databind.codegen.config.DefaultBindingConfiguration;
 import gov.nist.secauto.metaschema.databind.io.BindingException;
-import gov.nist.secauto.metaschema.databind.io.Format;
-import gov.nist.secauto.metaschema.databind.io.IDeserializer;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
 import org.junit.platform.commons.util.ReflectionUtils;
 
 import java.io.IOException;
-import java.io.Writer;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import edu.umd.cs.findbugs.annotations.NonNull;
-import edu.umd.cs.findbugs.annotations.Nullable;
-
-class BasicMetaschemaTest {
-  private static final ModuleLoader LOADER = new ModuleLoader();
-  private static final Logger LOGGER = LogManager.getLogger(BasicMetaschemaTest.class);
-  @TempDir
-  Path generationDir;
-  // @NonNull
-  // Path generationDir =
-  // ObjectUtils.notNull(Paths.get("target/generated-test-sources/metaschema"));
-
-  @NonNull
-  private static IModule loadModule(@NonNull Path moduleFile) throws MetaschemaException, IOException {
-    return LOADER.load(moduleFile);
-  }
-
-  public static Class<?> compileModule(@NonNull Path moduleFile, @Nullable Path bindingFile,
-      @NonNull String rootClassName, @NonNull Path classDir)
-      throws IOException, ClassNotFoundException, MetaschemaException {
-    IModule module = loadModule(moduleFile);
-
-    DefaultBindingConfiguration bindingConfiguration = new DefaultBindingConfiguration();
-    if (bindingFile != null && Files.exists(bindingFile) && Files.isRegularFile(bindingFile)) {
-      bindingConfiguration.load(bindingFile);
-    }
-
-    ModuleCompilerHelper.compileModule(module, classDir, bindingConfiguration);
-
-    // Load classes
-    return ModuleCompilerHelper.newClassLoader(
-        classDir,
-        ObjectUtils.notNull(Thread.currentThread().getContextClassLoader()))
-        .loadClass(rootClassName);
-  }
-
-  private static Object read(@NonNull Format format, @NonNull Path file, @NonNull Class<?> rootClass)
-      throws IOException {
-    IBindingContext context = IBindingContext.instance();
-
-    IDeserializer<?> deserializer = context.newDeserializer(format, rootClass);
-    LOGGER.info("Reading content: {}", file);
-    Object value = deserializer.deserialize(file);
-    return value;
-  }
-
-  private static <CLASS> void write(@NonNull Format format, @NonNull Path file, CLASS rootObject) throws IOException {
-    IBindingContext context = IBindingContext.instance();
-    @SuppressWarnings("unchecked") Class<CLASS> clazz = (Class<CLASS>) rootObject.getClass();
-
-    try (Writer writer = Files.newBufferedWriter(file, StandardOpenOption.CREATE, StandardOpenOption.WRITE,
-        StandardOpenOption.TRUNCATE_EXISTING)) {
-      assert writer != null;
-      context.newSerializer(format, clazz).serialize(rootObject, writer);
-    }
-  }
-
-  private static void runTests(@NonNull String testPath, @NonNull String rootClassName, @NonNull Path classDir)
-      throws ClassNotFoundException, IOException, MetaschemaException, BindingException {
-    runTests(testPath, rootClassName, classDir, null);
-  }
-
-  @SuppressWarnings("unused")
-  private static void runTests(
-      @NonNull String testPath,
-      @NonNull String rootClassName,
-      @NonNull Path classDir,
-      java.util.function.Consumer<Object> assertions)
-      throws ClassNotFoundException, IOException, MetaschemaException, BindingException {
-
-    Class<?> rootClass = compileModule(
-        ObjectUtils.notNull(Paths.get(String.format("src/test/resources/metaschema/%s/metaschema.xml", testPath))),
-        Paths.get(String.format("src/test/resources/metaschema/%s/binding.xml", testPath)),
-        rootClassName,
-        classDir);
-
-    assert rootClass != null;
-
-    Path xmlExample = Paths.get(String.format("src/test/resources/metaschema/%s/example.xml",
-        testPath));
-    LOGGER.info("Testing XML file: {}", xmlExample.toString());
-    if (Files.exists(xmlExample)) {
-      String xml;
-      {
-        Object root = read(Format.XML, xmlExample, rootClass);
-        LOGGER.atDebug().log("Read XML: Object: {}", root.toString());
-        if (assertions != null) {
-          assertAll("Deserialize XML", () -> {
-            assertions.accept(root);
-          });
-        }
-
-        LOGGER.atDebug().log("Write XML:");
-        write(Format.XML, ObjectUtils.notNull(Paths.get("target/out.xml")), root);
-
-        LOGGER.atDebug().log("Write JSON:");
-        write(Format.XML, ObjectUtils.notNull(Paths.get("target/out.json")), root);
-      }
-
-      Object root = read(Format.XML, ObjectUtils.notNull(Paths.get("target/out.xml")), rootClass);
-      if (assertions != null) {
-        assertAll("Deserialize XML (roundtrip)", () -> assertions.accept(root));
-      }
-    }
-  }
+class BasicMetaschemaTest
+    extends AbstractMetaschemaTest {
 
   @Test
   void testSimpleMetaschema() throws MetaschemaException, IOException, ClassNotFoundException, BindingException {
